Added SizeFactor property
Added a SizeFactor property which allows you to scale up the width of the barcode by specifying a multiplier
Default = 1
For example, specifying a SizeFactor of 2 basically doubles the width of the barcode. You can specify floating point values too. 
Tested final result with a Citizen CL-S321 Printer and Wasp WLS9600 scanner, with a value of 1.5. Worked well.
This actually makes scanning the barcode more reliable too, due to larger bar size, which works great on lower resolution barcode printers.